Rename dispatchers -> routers
